Separate out some specific workshop classes.
The workshop template is mostly plain bootstrap, so getting it work
alongside the site CSS seems mostly straightforward. We only need to
add in a .workshop class to the body, and a .card class to the
.container in the workshop. Everything else is bootstrap content.

This commit also removes the banner.html include since this did not
seem to be removed in the main site, but was referenced in the CSS.
Removing it makes it clear the .banner class is only needed in the
workshop, not here.