Explain R packages instead of database connections
While it is no doubt a useful feature of dplyr to "work directly with data stored in an external database", I think the discussion of this feature is more confusing than helpful at this point in the lesson (especially since it is never returned to or practiced later on). My suggestion (and what I've done here) is to spend a few moments discussing the concept of an R package using dplyr and tidyr as examples. This would arm beginning learners with the knowledge that R packages probably exist to support whatever their needs are in R so that they can seek them out as needed. 

I think it would be helpful to see the relational database portion expanded into another episode that walks the learner through the process of accessing and analyzing data from an external database using dplyr.

Also, I have changed the data frame referenced in line 96 (now line 92) from "surveys" to "interviews", the named data frame in the lesson.